Fix windows warnings

+++ kstars/dialogs/catalogobjectlistmodel.h
@@ -10,33 +10,33 @@
 
 class CatalogObjectListModel : public QAbstractTableModel
 {
-    Q_OBJECT
-  public:
-    CatalogObjectListModel(QObject *parent                    = nullptr,
-                           std::vector<CatalogObject> objects = {});
+        Q_OBJECT
+    public:
+        CatalogObjectListModel(QObject *parent                    = nullptr,
+                               std::vector<CatalogObject> objects = {});
 
-    inline int rowCount(const QModelIndex & = QModelIndex()) const override
-    {
-        return m_objects.size();
-    };
+        inline int rowCount(const QModelIndex & = QModelIndex()) const override
+        {
+            return static_cast<int>(m_objects.size());
+        };
 
-    inline int columnCount(const QModelIndex & = QModelIndex()) const override
-    {
-        return 11;
-    };
+        inline int columnCount(const QModelIndex & = QModelIndex()) const override
+        {
+            return 11;
+        };
 
-    QVariant headerData(int section, Qt::Orientation orientation,
-                        int role = Qt::DisplayRole) const override;
+        QVariant headerData(int section, Qt::Orientation orientation,
+                            int role = Qt::DisplayRole) const override;
 
-    QVariant data(const QModelIndex &index, int role = Qt::DisplayRole) const override;
+        QVariant data(const QModelIndex &index, int role = Qt::DisplayRole) const override;
 
-    void setObjects(std::vector<CatalogObject> objects);
+        void setObjects(std::vector<CatalogObject> objects);
 
-    inline const CatalogObject &getObject(const QModelIndex &index) const
-    {
-        return m_objects.at(index.row());
-    };
+        inline const CatalogObject &getObject(const QModelIndex &index) const
+        {
+            return m_objects.at(index.row());
+        };
 
-  private:
-    std::vector<CatalogObject> m_objects;
+    private:
+        std::vector<CatalogObject> m_objects;
 };
